all_requirements.txt
# Core FastAPI dependencies
fastapi==0.108.0
uvicorn==0.25.0
pydantic-settings==2.1.0

# HTTP and networking
urllib3==1.26.19  # needed for OpenSSL 1.1.1+
requests==2.31.0

# Database adapters
pymongo==4.6.2          # MongoDB for api-game
psycopg2-binary==2.9.9  # PostgreSQL for api-site

# Real-time communication
websockets==12.0

# AWS and environment
boto3==1.34.0
python-dotenv==1.0.0

# Development and testing (optional)
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0  # for testing FastAPI apps 